CPLP-793: Merged validation, moved redux function to prop provided function, added TODOs
