+++ cx-portal/src/components/pages/TechnicalUserManagement/index.tsx
@@ -1,16 +1,30 @@
-import { useState } from 'react'
-import { PageHeader } from 'cx-portal-shared-components'
-import { useTranslation } from 'react-i18next'
-import { AddTechnicalUserOverlay } from './AddTechnicalUserOverlay'
+import './TechnicalUserManagement.scss'
+import { AddTechnicalUserOverlay, DefaultFormFieldValuesType } from './AddTechnicalUserOverlay'
+import { AddTechnicalUserResponseOverlay } from './AddTechnicalUserResponseOverlay'
 import { ContentTechnicalUser } from './ContentTechnicalUser'
 import { PageBreadcrumb } from 'components/shared/frame/PageBreadcrumb/PageBreadcrumb'
-import './TechnicalUserManagement.scss'
-import { AddTechnicalUserResponseOverlay } from './AddTechnicalUserResponseOverlay'
+import { PageHeader } from 'cx-portal-shared-components'
+import { RequestState } from 'types/MainTypes'
+import { addItem, fetchPage } from 'features/admin/service/actions'
+import { stateSelector as createSelector } from 'features/admin/service/screate'
+import { useDispatch, useSelector } from 'react-redux'
+import { useEffect, useState } from 'react'
+import { useTranslation } from 'react-i18next'
 
 export default function TechnicalUserManagement() {
   const [open, setOpen] = useState(false)
   const [openResponse, setOpenResponse] = useState(false)
   const { t } = useTranslation()
+
+  const dispatch = useDispatch()
+  // TODO: Check if creatResult data is still valid since form validation is done by React Hook Form
+  const createResult = useSelector(createSelector)
+  useEffect(() => {
+    //reload the data after successful create
+    if (createResult.request === RequestState.OK) {
+      dispatch(fetchPage(0))
+    }
+  }, [dispatch, createResult])
 
   const openAddTechnicalUserOverlay = () => {
     setOpen(true)
@@ -28,9 +42,19 @@ export default function TechnicalUserManagement() {
     setOpenResponse(false)
   }
 
-  const handleAddUTechnicalUserConfirm = () => {
-    // TODO: API call and make response open followup dialog with openAddTechnicalUserResponseOverlay()
+  const handleAddTechnicalUserConfirm = (formValues: DefaultFormFieldValuesType) => {
     openAddTechnicalUserResponseOverlay()
+
+    console.log('Form data: ', formValues)
+
+    // TODO: use above formValues for API call/Redux
+    // dispatch(
+    //     addItem({
+    //       name: `testaccount-${Date.now()}`,
+    //       description: 'another test account',
+    //       authenticationType: 'SECRET',
+    //     })
+    //   )
   }
 
   return (
@@ -38,7 +62,7 @@ export default function TechnicalUserManagement() {
       <AddTechnicalUserOverlay
         dialogOpen={open}
         handleClose={closeAddTechnicalUserOverlay}
-        handleConfirm={handleAddUTechnicalUserConfirm}
+        handleConfirm={handleAddTechnicalUserConfirm}
       />
       <AddTechnicalUserResponseOverlay
         dialogOpen={openResponse}
@@ -51,7 +75,6 @@ export default function TechnicalUserManagement() {
       >
         <PageBreadcrumb backButtonVariant="contained" />
       </PageHeader>
-
       <ContentTechnicalUser
         openAddTechnicalUserOverlay={openAddTechnicalUserOverlay}
       />
